Add support for `code` command's use inside tmux + name as arg + auto vim
Beforehand it would just fail if `code` was run inside of a tmux
session. Now it will switch to the session chosen if it exists, or
otherwise create the session and switch to it.

Now one can specifiy the project name as an argument, i.e.
`code project-name`. An exact match will attach or create and attach to
the tmux session matching the project name. Otherwise an fzf fuzzy
search UI will appear to refine the search. If exactly one result
matches the name, fzf will automatically and transparently select that
project name.

Vim will be automatically launched when a new session is created. Care
has been taken to ensure that closing vim will not close the session
window, as is the common behaviour.