Add vim-tmux-navigator + fix broken tmux flags
vim-tmux-navigator allows vim and tmux to use the same keymaps to
navigate splits (control h/j/k/l by default). It works in conjunction
with a corresponding neovim plugin.

tmux flags such as `-V` (version) were broken due to my nix technique
of wrapping the executable with custom arguments. My use of `source` was
invalidating tmux args which were not compatible with it. The solution
was to run `tmux source` as a separate invocation before the main
invocation of tmux which uses the original user supplied args.